requirements-dev.txt
# Development Dependencies
-r requirements.txt

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0  # For async HTTP requests in tests

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
types-python-dateutil==2.8.19.14

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==2.0.0

# Development Tools
ipython==8.20.0
jupyter==1.0.0
pre-commit==3.6.0

# Debugging
pdbpp==0.10.3
ipdb==0.13.13

# Test database
pytest-postgresql==4.1.0
pytest-redis==2.2.0

# Mocking
respx==0.20.2

# Code coverage
coverage==7.4.0
codecov==2.1.13

# Security
safety==2.3.5
bandit==1.7.6

# Type checking
sqlalchemy2-stubs==0.0.2a35
types-PyJWT==2.8.0.20240218
types-python-dateutil==2.8.19.14
types-requests==2.31.0.20240125
types-pyyaml==6.0.12.12

# Development server
uvicorn[standard]==0.27.0
watchfiles==0.21.0  # For auto-reload
